PS5-system folder restructuring, pytest configuration
