+++ data/pairs.txt
@@ -1,0 +1,4 @@
+1,9,5,0,20,-4,12,16,7 12
+1,2,3 5
+a,b,c 123
+1,2,3,5,4
