--- pairs.txt
@@ -1,1 +1,0 @@
-1,9,5,0,20,-4,12,16,7 12
